common lib name change and refactor
