+++ apps/backend/gateway/src/auth/auth.service.ts
@@ -1,7 +1,7 @@
 import { AuthLoginPayload, CompleteSetupPayload, USER_PATTERNS, USER_SERVICE } from '@limbo/users-contracts';
 import { Inject, Injectable } from '@nestjs/common';
 import { ClientProxy } from '@nestjs/microservices';
-import { CompleteSetupDto } from '../users/dtos/complete-setup.dto';
+import { CompleteSetupDto } from '../auth/dtos/complete-setup.dto';
 
 @Injectable()
 export class AuthService {
