--- libs/common/src/lib/Users/index.ts
@@ -1,3 +1,0 @@
-export * from './base';
-export * from './request';
-export * from './response';
