--- libs/common/src/lib/Users/request/index.ts
@@ -1,3 +1,0 @@
-export * from './admin-create-user.request.interface';
-export * from './complete-setup.request';
-export * from './login.request.interface';
