+++ libs/common/src/lib/auth/index.ts
@@ -1,2 +1,4 @@
+export * from './complete-setup.request';
+export * from './login.request';
 export * from './login.response.interface';
 export * from './pending-login.response.interface';
