+++ libs/common/src/lib/auth/pending-login.response.interface.ts
