+++ libs/common/src/lib/index.ts
@@ -1,1 +1,2 @@
-export * from './Users';
+export * from './auth';
+export * from './users';
