+++ libs/common/src/lib/users/index.ts
@@ -1,3 +1,4 @@
+export * from './admin-create-user.request';
 export * from './user-role.enum';
 export * from './user-status.enum';
 export * from './user.dto';
